+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## Version
+## Unreleased
 
-`Date`
+`2021-06-XX`
 
 ### What's new
 
+- **Added expanding functionality for `Table`.** Provide `subComponent` prop.
 - **Improved `Table` documentation.**
 
 ## [1.9.0]